refactor: 공지사항 관련 페이지 컴포넌트 트리 구조 정리
