+++ src/views/announcement/ui/index.ts
@@ -1,9 +1,8 @@
-export * from './announcement-management-page'
-export * from './announcement-management-search-page'
-export * from './create-announcement-form-page'
-export * from './announcement-management-detail-page'
-export * from './modify-announcement-form-page'
-
+export * from './announcement-management-page-container'
 export * from './announcement-management-page-view'
+export * from './announcement-management-detail-page-container'
 export * from './announcement-management-detail-page-view'
+export * from './announcement-management-search-page-container'
 export * from './announcement-management-search-page-view'
+export * from './create-announcement-form-page'
+export * from './modify-announcement-form-page'
